requirements.txt
# === API Framework ===
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv

# === LangChain Core ===
langchain
langchain-community
langchain-openai
langchain-chroma

# === Embeddings ===
langchain-huggingface
sentence-transformers

# === Vector Store ===
chromadb

# === Database ===
sqlalchemy
psycopg2-binary

# === LLM Providers ===
openai
tiktoken

# === Agent Orchestration ===
langgraph

# === Utilities ===
numpy